+++ src/models/outgoing_webhook.rs
@@ -45,7 +45,7 @@ pub struct OutgoingWebhook {
     pub trigger_words: Option<Vec<String>>,
     /// When to trigger the webhook, `0` when a trigger word is present at all and `1` if the message starts with a trigger word
     #[serde(rename = "trigger_when", skip_serializing_if = "Option::is_none")]
-    pub trigger_when: Option<i64>,
+    pub trigger_when: Option<i32>,
     /// The URLs to POST the payloads to when the webhook is triggered
     #[serde(rename = "callback_urls", skip_serializing_if = "Option::is_none")]
     pub callback_urls: Option<Vec<String>>,
