+++ src/models/post_metadata_images_inner.rs
@@ -14,9 +14,9 @@
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct PostMetadataImagesInner {
     #[serde(rename = "height", skip_serializing_if = "Option::is_none")]
-    pub height: Option<i64>,
+    pub height: Option<i32>,
     #[serde(rename = "width", skip_serializing_if = "Option::is_none")]
-    pub width: Option<i64>,
+    pub width: Option<i32>,
 }
 
 impl PostMetadataImagesInner {
