+++ src/models/subscription_stats.rs
@@ -14,7 +14,7 @@
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct SubscriptionStats {
     #[serde(rename = "remaining_seats", skip_serializing_if = "Option::is_none")]
-    pub remaining_seats: Option<i64>,
+    pub remaining_seats: Option<i32>,
     #[serde(rename = "is_paid_tier", skip_serializing_if = "Option::is_none")]
     pub is_paid_tier: Option<String>,
 }
