+++ src/models/team_stats.rs
@@ -16,9 +16,9 @@ pub struct TeamStats {
     #[serde(rename = "team_id", skip_serializing_if = "Option::is_none")]
     pub team_id: Option<String>,
     #[serde(rename = "total_member_count", skip_serializing_if = "Option::is_none")]
-    pub total_member_count: Option<i64>,
+    pub total_member_count: Option<i32>,
     #[serde(rename = "active_member_count", skip_serializing_if = "Option::is_none")]
-    pub active_member_count: Option<i64>,
+    pub active_member_count: Option<i32>,
 }
 
 impl TeamStats {
